[FIX] unsigned variables

+++ 0x14-bit_manipulation/5-flip_bits.c
@@ -7,7 +7,7 @@
  */
 unsigned int flip_bits(unsigned long int n, unsigned long int m)
 {
-	int bits,  i = 0;
+	unsigned long int bits,  i = 0;
 
 	bits = n ^ m;
